+++ src/components/InstancesList/InstancesList.js
@@ -40,6 +40,7 @@ import {
   parseFiltersToStr,
   marshalInstance,
   omitFromArray,
+  isTestEnv,
 } from '../../utils';
 import {
   INSTANCES_ID_REPORT_TIMEOUT,
@@ -51,6 +52,7 @@ import {
 } from '../../reports';
 import ErrorModal from '../ErrorModal';
 import CheckboxColumn from './CheckboxColumn';
+import SelectedRecordsModal from '../SelectedRecordsModal';
 
 import { buildQuery } from '../../routes/buildManifestObject';
 
@@ -59,7 +61,7 @@ import css from './instances.css';
 const INITIAL_RESULT_COUNT = 30;
 const RESULT_COUNT_INCREMENT = 30;
 
-class InstancesView extends React.Component {
+class InstancesList extends React.Component {
   static defaultProps = {
     browseOnly: false,
     showSingleResult: true,
@@ -93,8 +95,10 @@ class InstancesView extends React.Component {
     showNewFastAddModal: false,
     inTransitItemsExportInProgress: false,
     instancesIdExportInProgress: false,
+    instancesQuickExportInProgress: false,
     showErrorModal: false,
     selectedRows: {},
+    isSelectedRecordsModalOpened: false,
   };
 
   onFilterChangeHandler = ({ name, values }) => {
@@ -222,10 +226,10 @@ class InstancesView extends React.Component {
 
         clearTimeout(infoCalloutTimer);
 
-        const report = new InstancesIdReport();
+        const report = new InstancesIdReport('SearchInstanceUUIDs');
 
         if (!isEmpty(items)) {
-          report.toCSV(items);
+          report.toCSV(items, record => record.id);
         }
       } catch (error) {
         clearTimeout(infoCalloutTimer);
@@ -240,8 +244,34 @@ class InstancesView extends React.Component {
     });
   };
 
+  triggerQuickExport = async (sendCallout) => {
+    const { instancesQuickExportInProgress } = this.state;
+
+    if (instancesQuickExportInProgress) return;
+
+    this.setState({ instancesQuickExportInProgress: true });
+
+    try {
+      const instanceIds = Object.keys(this.state.selectedRows).slice(0, QUICK_EXPORT_LIMIT);
+
+      await this.props.parentMutator.quickExport.POST({
+        uuids: instanceIds,
+        type: 'uuid',
+        recordType: 'INSTANCE'
+      });
+      new InstancesIdReport('QuickInstanceExport').toCSV(instanceIds);
+    } catch (error) {
+      sendCallout({
+        type: 'error',
+        message: <FormattedMessage id="ui-inventory.communicationProblem" />,
+      });
+    } finally {
+      this.setState({ instancesQuickExportInProgress: false });
+    }
+  };
+
   generateCQLQueryReport = async () => {
-    if (process.env.NODE_ENV !== 'test') {
+    if (!isTestEnv()) {
       const { data } = this.props;
 
       const query = buildQuery(data.query, {}, data, { log: noop }, this.props);
@@ -343,7 +373,7 @@ class InstancesView extends React.Component {
           id: 'dropdown-clickable-export-marc',
           icon: 'download',
           messageId: 'ui-inventory.exportInstancesInMARC',
-          onClickHandler: buildOnClickHandler(noop),
+          onClickHandler: buildOnClickHandler(this.triggerQuickExport),
           isDisabled: !selectedRowsCount || isQuickExportLimitExceeded,
         })}
         {isQuickExportLimitExceeded && (
@@ -364,6 +394,13 @@ class InstancesView extends React.Component {
           onClickHandler: buildOnClickHandler(noop),
           isDisabled: true,
         })}
+        {this.getActionItem({
+          id: 'dropdown-clickable-show-selected-records',
+          icon: 'eye-open',
+          messageId: 'ui-inventory.instances.showSelectedRecords',
+          onClickHandler: buildOnClickHandler(() => this.setState({ isSelectedRecordsModalOpened: true })),
+          isDisabled: !selectedRowsCount,
+        })}
       </>
     );
   };
@@ -372,7 +409,10 @@ class InstancesView extends React.Component {
     this.setState({ showErrorModal: false });
   };
 
-  toggleRowSelection = rowId => {
+  toggleRowSelection = ({
+    id: rowId,
+    ...rowData
+  }) => {
     this.setState(({ selectedRows }) => {
       const isRowSelected = Boolean(selectedRows[rowId]);
       const newSelectedRows = { ...selectedRows };
@@ -380,7 +420,7 @@ class InstancesView extends React.Component {
       if (isRowSelected) {
         delete newSelectedRows[rowId];
       } else {
-        newSelectedRows[rowId] = true;
+        newSelectedRows[rowId] = rowData;
       }
 
       return { selectedRows: newSelectedRows };
@@ -389,6 +429,10 @@ class InstancesView extends React.Component {
 
   handleResetAll = () => {
     this.setState({ selectedRows: {} });
+  }
+
+  handleSelectedRecordsModalCancel = () => {
+    this.setState({ isSelectedRecordsModalOpened: false });
   }
 
   renderPaneSub() {
@@ -425,14 +469,24 @@ class InstancesView extends React.Component {
         path,
       }
     } = this.props;
+    const {
+      isSelectedRecordsModalOpened,
+      selectedRows,
+    } = this.state;
 
     const resultsFormatter = {
-      'select': ({ id }) => (
+      'select': ({
+        id,
+        ...rowData
+      }) => (
         <CheckboxColumn>
           <Checkbox
-            checked={Boolean(this.state.selectedRows[id])}
+            checked={Boolean(selectedRows[id])}
             aria-label={intl.formatMessage({ id: 'ui-inventory.instances.rows.select' })}
-            onChange={() => this.toggleRowSelection(id)}
+            onChange={() => this.toggleRowSelection({
+              id,
+              ...rowData
+            })}
           />
         </CheckboxColumn>
       ),
@@ -450,6 +504,14 @@ class InstancesView extends React.Component {
       'publishers': r => r.publication.map(p => (p ? `${p.publisher} ${p.dateOfPublication ? `(${p.dateOfPublication})` : ''}` : '')).join(', '),
       'publication date': r => r.publication.map(p => p.dateOfPublication).join(', '),
       'contributors': r => formatters.contributorsFormatter(r, data.contributorTypes),
+    };
+
+    const columnMapping = {
+      select: '',
+      title: intl.formatMessage({ id: 'ui-inventory.instances.columns.title' }),
+      contributors: intl.formatMessage({ id: 'ui-inventory.instances.columns.contributors' }),
+      publishers: intl.formatMessage({ id: 'ui-inventory.instances.columns.publishers' }),
+      relation: intl.formatMessage({ id: 'ui-inventory.instances.columns.relation' }),
     };
 
     const formattedSearchableIndexes = searchableIndexes.map(index => {
@@ -482,13 +544,7 @@ class InstancesView extends React.Component {
               source: 'FOLIO',
             }}
             visibleColumns={visibleColumns || ['select', 'title', 'contributors', 'publishers', 'relation']}
-            columnMapping={{
-              select: '',
-              title: intl.formatMessage({ id: 'ui-inventory.instances.columns.title' }),
-              contributors: intl.formatMessage({ id: 'ui-inventory.instances.columns.contributors' }),
-              publishers: intl.formatMessage({ id: 'ui-inventory.instances.columns.publishers' }),
-              relation: intl.formatMessage({ id: 'ui-inventory.instances.columns.relation' }),
-            }}
+            columnMapping={columnMapping}
             columnWidths={{
               select: '30px',
               title: '40%',
@@ -526,6 +582,13 @@ class InstancesView extends React.Component {
           content={<FormattedMessage id="ui-inventory.reports.inTransitItem.emptyReport.message" />}
           onClose={this.onErrorModalClose}
         />
+        <SelectedRecordsModal
+          isOpen={isSelectedRecordsModalOpened}
+          selectedRecords={selectedRows}
+          columnMapping={columnMapping}
+          formatter={resultsFormatter}
+          onCancel={this.handleSelectedRecordsModalCancel}
+        />
       </>
     );
   }
@@ -534,4 +597,4 @@ class InstancesView extends React.Component {
 export default flowRight(
   injectIntl,
   withLocation,
-)(InstancesView);
+)(InstancesList);