+++ test/fixtures/instances.js
@@ -1,0 +1,225 @@
+export const instances = [
+  {
+    'id': '3afe2364-c584-46b7-8699-9667d4244d35',
+    'hrid': 'in00000000001',
+    'source': 'MARC',
+    'title': '#youthaction',
+    'indexTitle': '#youthaction : becoming political in the digital age /',
+    'alternativeTitles': [{
+      'alternativeTitleTypeId': 'fe19bae4-da28-472b-be90-d442e2428ead',
+      'alternativeTitle': 'Hashtag youthaction'
+    }, {
+      'alternativeTitleTypeId': 'fe19bae4-da28-472b-be90-d442e2428ead',
+      'alternativeTitle': 'Youthaction'
+    }],
+    'editions': [],
+    'series': [],
+    'identifiers': [{
+      'identifierTypeId': '7e591197-f335-4afb-bc6d-a6d76ca3bace',
+      'value': '18398974'
+    }, {
+      'identifierTypeId': 'c858e4f2-2b6b-4385-842b-60732ee14abb',
+      'value': '2014030023'
+    }, {
+      'identifierTypeId': '8261054f-be78-422d-bd51-4ed9f33c3422',
+      'value': '9781623967956 (pbk.)'
+    }, {
+      'identifierTypeId': '8261054f-be78-422d-bd51-4ed9f33c3422',
+      'value': '9781623967963 (hardcover)'
+    }, {
+      'identifierTypeId': 'fcca2643-406a-482a-b760-7a7f8aec640e',
+      'value': '9781623967970 (ebook)'
+    }],
+    'contributors': [{
+      'contributorNameTypeId': '2b94c631-fca9-4892-a730-03ee529ffe2a',
+      'name': 'Kirshner, Benjamin',
+      'contributorTypeId': null,
+      'contributorTypeText': null,
+      'primary': false
+    }, {
+      'contributorNameTypeId': '2b94c631-fca9-4892-a730-03ee529ffe2a',
+      'name': 'Middaugh, Ellen',
+      'contributorTypeId': null,
+      'contributorTypeText': null,
+      'primary': false
+    }],
+    'subjects': ['Information society--Political aspects', 'Social media--Political aspects', 'Political participation--Technological innovations', 'Youth--Political activity'],
+    'classifications': [{
+      'classificationNumber': 'HM851 .A112 2015',
+      'classificationTypeId': 'ce176ace-a53e-4b4d-aa89-725ed7b2edac'
+    }, {
+      'classificationNumber': '302.231',
+      'classificationTypeId': '42471af9-7d25-4f3a-bf78-60d29dcf463b'
+    }],
+    'publication': [{
+      'publisher': 'Information Age Publishing, Inc.',
+      'place': 'Charlotte, NC',
+      'dateOfPublication': '2015',
+      'role': 'Publication'
+    }],
+    'publicationFrequency': [],
+    'publicationRange': [],
+    'electronicAccess': [],
+    'instanceTypeId': '6312d172-f0cf-40f6-b27d-9fa8feaf332f',
+    'instanceFormatIds': ['8d511d33-5e85-4c5d-9bce-6e3c9cd0c324'],
+    'physicalDescriptions': ['ix, 217 pages ; 24 cm.'],
+    'languages': ['eng'],
+    'notes': [{
+      'instanceNoteTypeId': '86b6e817-e1bc-42fb-bab0-70e7547de6c1',
+      'note': 'Includes bibliographical references',
+      'staffOnly': false
+    }, {
+      'instanceNoteTypeId': '5ba8e385-0e27-462e-a571-ffa1fa34ea54',
+      'note': "Educating powerful citizens in a changing world / Ellen Middaugh and Ben Kirshner -- And the weapon of choice was-- / McCrae Adams Parker -- Phones aren't smart until you tell them what to do / Elisabeth Soep -- Leveraging new media and youths : repertoires of practice toward connected learning / Kris D. Gutierrez, Althea Scott Nixon, and Jevon D. Hunter -- Compugirls speak : how we use social media for social movements / Mitzi Vilchis, Kimberly A. Scott, and Courtney Besaw -- \"Son enough\" : developing girls' agency in india across domestic and civic spheres through feminist media practice / Glynda Hull, Mark Jury, and Urvashi Sahni -- Lost, sweaty, and engaged in dialogue : the civic opportunities of geospatial play / Antero Garcia and Ellen Middaugh -- Dialogue across difference : a case study of facing history and ourselves : digital media innovation networ / Justin Reich, Anna L. Romer, and Dennis Barr -- Blogging as civic engagement : developing a sense of authority and audience in an urban public school classroom / Katherine Schultz, Erica Hodgin, and Johanna Paraiso -- The making of a global citizen : a model of supporting civic learning opportunities among urban Latino youth / Brendesha M. Tynes and Vanessa Monterosa -- Politically engaged and alienated youth : reevaluating 2010 UK student protests / Allen L. Linton -- Sustaining innovation : lessons for digitally mediated civic education / Ellen Middaugh",
+      'staffOnly': false
+    }],
+    'modeOfIssuanceId': '9d18a02f-5897-4c31-9106-c9abb5c7ae8b',
+    'discoverySuppress': false,
+    'statisticalCodeIds': [],
+    'statusUpdatedDate': '2020-12-24T13:14:31.537+0000',
+    'metadata': {
+      'createdDate': '2020-12-24T13:14:31.538+00:00',
+      'createdByUserId': '2bf7485a-4fdb-5cd8-ae39-dbac373e78fa',
+      'updatedDate': '2020-12-24T13:36:13.487+00:00',
+      'updatedByUserId': '2bf7485a-4fdb-5cd8-ae39-dbac373e78fa'
+    },
+    'tags': {
+      'tagList': []
+    },
+    'natureOfContentTermIds': [],
+    'links': {
+      'self': 'http://folio-testing-okapi.dev.folio.org/inventory/instances/3afe2364-c584-46b7-8699-9667d4244d35'
+    }
+  },
+  {
+    'id': '5bf370e0-8cca-4d9c-82e4-5170ab2a0a39',
+    'hrid': 'inst000000000022',
+    'source': 'FOLIO',
+    'title': 'A semantic web primer',
+    'indexTitle': 'Semantic web primer',
+    'alternativeTitles': [],
+    'editions': [],
+    'series': ['Cooperative information systems'],
+    'identifiers': [{
+      'identifierTypeId': '8261054f-be78-422d-bd51-4ed9f33c3422',
+      'value': '0262012103'
+    }, {
+      'identifierTypeId': '8261054f-be78-422d-bd51-4ed9f33c3422',
+      'value': '9780262012102'
+    }, {
+      'identifierTypeId': 'c858e4f2-2b6b-4385-842b-60732ee14abb',
+      'value': '2003065165'
+    }],
+    'contributors': [{
+      'contributorNameTypeId': '2b94c631-fca9-4892-a730-03ee529ffe2a',
+      'name': 'Antoniou, Grigoris',
+      'contributorTypeId': '6e09d47d-95e2-4d8a-831b-f777b8ef6d81',
+      'contributorTypeText': '',
+      'primary': null
+    }, {
+      'contributorNameTypeId': '2b94c631-fca9-4892-a730-03ee529ffe2a',
+      'name': 'Van Harmelen, Frank',
+      'contributorTypeId': '6e09d47d-95e2-4d8a-831b-f777b8ef6d81',
+      'contributorTypeText': '',
+      'primary': null
+    }],
+    'subjects': ['Semantic Web'],
+    'classifications': [{
+      'classificationNumber': '025.04',
+      'classificationTypeId': '42471af9-7d25-4f3a-bf78-60d29dcf463b'
+    }, {
+      'classificationNumber': 'TK5105.88815 .A58 2004',
+      'classificationTypeId': 'ce176ace-a53e-4b4d-aa89-725ed7b2edac'
+    }],
+    'publication': [{
+      'publisher': 'MIT Press',
+      'place': 'Cambridge, Mass. ',
+      'dateOfPublication': 'c2004',
+      'role': 'Publisher'
+    }],
+    'publicationFrequency': [],
+    'publicationRange': [],
+    'electronicAccess': [],
+    'instanceTypeId': '6312d172-f0cf-40f6-b27d-9fa8feaf332f',
+    'instanceFormatIds': [],
+    'physicalDescriptions': ['xx, 238 p. : ill. ; 24 cm.'],
+    'languages': ['eng'],
+    'notes': [{
+      'instanceNoteTypeId': '6a2533a7-4de2-4e64-8466-074c2fa9308c',
+      'note': 'Includes bibliographical references and index.',
+      'staffOnly': false
+    }, {
+      'instanceNoteTypeId': '6a2533a7-4de2-4e64-8466-074c2fa9308c',
+      'note': 'The development of the Semantic Web, with machine-readable content, has the potential to revolutionize the World Wide Web and its uses. A Semantic Web Primer provides an introduction and guide to this continuously evolving field, describing its key ideas, languages, and technologies. Suitable for use as a textbook or for independent study by professionals, it concentrates on undergraduate-level fundamental concepts and techniques that will enable readers to proceed with building applications on their own and includes exercises, project descriptions, and annotated references to relevant online materials. The third edition of this widely used text has been thoroughly updated, with significant new material that reflects a rapidly developing field. Treatment of the different languages (OWL2, rules) expands the coverage of RDF and OWL, defining the data model independently of XML and including coverage of N3/Turtle and RDFa. A chapter is devoted to OWL2, the new W3C standard. This edition also features additional coverage of the query language SPARQL, the rule language RIF and the possibility of interaction between rules and ontology languages and applications. The chapter on Semantic Web applications reflects the rapid developments of the past few years. A new chapter offers ideas for term projects',
+      'staffOnly': false
+    }],
+    'modeOfIssuanceId': '9d18a02f-5897-4c31-9106-c9abb5c7ae8b',
+    'catalogedDate': '2019-04-05',
+    'discoverySuppress': false,
+    'statisticalCodeIds': ['b5968c9e-cddc-4576-99e3-8e60aed8b0dd'],
+    'statusId': '9634a5ab-9228-4703-baf2-4d12ebc77d56',
+    'statusUpdatedDate': '2020-12-24T03:18:49.069+0000',
+    'metadata': {
+      'createdDate': '2020-12-24T03:18:49.069+00:00',
+      'createdByUserId': null,
+      'updatedDate': '2020-12-24T03:18:49.069+00:00',
+      'updatedByUserId': null
+    },
+    'tags': {
+      'tagList': []
+    },
+    'natureOfContentTermIds': [],
+    'links': {
+      'self': 'http://folio-testing-okapi.dev.folio.org/inventory/instances/5bf370e0-8cca-4d9c-82e4-5170ab2a0a39'
+    }
+  },
+  {
+    'id': '69640328-788e-43fc-9c3c-af39e243f3b7',
+    'hrid': 'inst000000000001',
+    'source': 'FOLIO',
+    'title': 'ABA Journal',
+    'alternativeTitles': [],
+    'editions': [],
+    'series': [],
+    'identifiers': [{
+      'identifierTypeId': '913300b2-03ed-469a-8179-c1092c991227',
+      'value': '0747-0088'
+    }, {
+      'identifierTypeId': 'c858e4f2-2b6b-4385-842b-60732ee14abb',
+      'value': '84641839'
+    }],
+    'contributors': [],
+    'subjects': [],
+    'classifications': [],
+    'publication': [{
+      'publisher': 'American Bar Association',
+      'place': 'Chicago, Ill.',
+      'dateOfPublication': '1915-1983',
+      'role': null
+    }],
+    'publicationFrequency': [],
+    'publicationRange': [],
+    'electronicAccess': [],
+    'instanceTypeId': '6312d172-f0cf-40f6-b27d-9fa8feaf332f',
+    'instanceFormatIds': [],
+    'physicalDescriptions': [],
+    'languages': [],
+    'notes': [],
+    'discoverySuppress': false,
+    'statisticalCodeIds': [],
+    'statusUpdatedDate': '2020-12-24T03:18:49.266+0000',
+    'metadata': {
+      'createdDate': '2020-12-24T03:18:49.266+00:00',
+      'createdByUserId': null,
+      'updatedDate': '2020-12-24T03:18:49.266+00:00',
+      'updatedByUserId': null
+    },
+    'tags': {
+      'tagList': []
+    },
+    'natureOfContentTermIds': ['0abeee3d-8ad2-4b04-92ff-221b4fce1075'],
+    'links': {
+      'self': 'http://folio-testing-okapi.dev.folio.org/inventory/instances/69640328-788e-43fc-9c3c-af39e243f3b7'
+    }
+  },
+];