+++ test/jest/__mock__/documentCreateRange.mock.js
@@ -1,0 +1,8 @@
+global.document.createRange = () => ({
+  setStart: () => {},
+  setEnd: () => {},
+  commonAncestorContainer: {
+    nodeName: 'BODY',
+    ownerDocument: document,
+  },
+});