+++ CHANGELOG.md
@@ -34,6 +34,10 @@
 * Upgraded to create-inventory plugin v2.0.0.
 * Include missing item-status values. Refs UIIN-1385.
 * Disable sorting by `relation` column. Refs UIIN-1303.
+* Add show selected records action menu. Refs UIIN-1375.
+* Implement quick export of instances in MARC Bib output format. Refs UIIN-1382.
+* Add jest setup. Refs UIIN-1375.
+* Cover select instances functionality with jest/rtl tests. Refs UIIN-1390.
 
 ## [5.0.1](https://github.com/folio-org/ui-inventory/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v5.0.0...v5.0.1)